+++ cognite/neat/rules/issues/importing.py
@@ -16,6 +16,7 @@ __all__ = [
     "MultipleDataModelsWarning",
     "UnknownPropertyTypeWarning",
     "FailedToInferValueTypeWarning",
+    "MoreThanOneNonAlphanumericCharacterWarning",
     "UnknownContainerConstraintWarning",
     "NoDataModelError",
     "ModelImportError",
@@ -273,6 +274,22 @@ class FailedImportWarning(ModelImportWarning):
 
     def dump(self) -> dict[str, Any]:
         return {"identifier": list(self.identifier)}
+
+
+@dataclass(frozen=True)
+class MoreThanOneNonAlphanumericCharacterWarning(ModelImportWarning):
+    description = """This warning is raised when doing regex validation of strings which either represent class ids,
+    property ids, prefix, data model name, that contain more than one non-alphanumeric character,
+    such as for example '_' or '-'."""
+
+    field_name: str
+    value: str
+
+    def message(self) -> str:
+        return f"Field {self.field_name} with value {self.value} contains more than one non-alphanumeric character!"
+
+    def dump(self) -> dict[str, str]:
+        return {"field_name": self.field_name, "value": self.value}
 
 
 @dataclass(frozen=True)