+++ docs/CHANGELOG.md
@@ -15,6 +15,32 @@ Changes are grouped as follows:
 - `Fixed` for any bug fixes.
 - `Security` in case of vulnerabilities.
 
+## [0.77.8] - 23-05-24
+### Fixed
+- In the conversion form Information to DMS Rules, when referencing a class in reference rules, the implements
+  was not set correctly. This is now fixed.
+- In the new implementation of the conversion between Information and DMS rules, containers that already exist
+  in a last or reference rule object were recreated. This is now fixed.
+
+## [0.77.7] - 23-05-24
+### Fixed
+- In the `DMSImporter`, if you imported a data model with multiple views referencing the same direct property
+  in a container, it would return an error. This is allowed and thus no longer return an error.
+- There was an edge case that could cause the conversion between Information and DMS rules to fail with
+  `MissingContainerError`. The conversion is now reimplemented to ensure that Information rules always
+  will create the necessary containers in the conversion to DMS rules.
+
+## [0.77.6] - 23-05-24
+### Improves
+- Documentation on how to use raw filter
+- Added a simple example of Rules with raw filter
+- Added new test for raw filter
+
+## [0.77.5] - 23-05-24
+### Fixed
+- `DMSExporter` creates the schema depending on `extension` in metadata field as defined in the
+  [documentation](https://cognite-neat.readthedocs-hosted.com/en/latest/terminology/rules-excel-input.html).
+
 ## [0.77.4] - 22-05-24
 ### Improves
 - Information rules are now read using InformationRulesInput data class, replicate the form of DMS rules.